checkpoint
Summary:
all parser.rs functions converted to methods of struct Parser

Add some preprocessing to inject block scope start/end

Update preprocessor to handle start of new top level scope, refactor the parser consumers to be aware of the new ParserResult type.

Change the Ast elements to use copy and clone semantics instead of the zero copy Tk and TkSlice stuff that the lexer uses.

Add type names to debug_locals output in interactive mode, move imports to top of file

Add prelim AST support for functions, returns, and blocks.

Note that we only allow functions with no arguments for now.

Customize json serialization for OwnedTk to ignore tag and serialize the byte vector as a string

Add positional arguments to func defs and calls, add a few more builtins for testing

Hacky as shit, but user defined functions seem to be working on a global namespace. They totally nuke the value stack because there are not frames and stuff Things to think about

Assignments from returnvalues of user defined functions are kind of wonky but the code for milestone v0.4.0 basically runs

Test Plan: N/a

Reviewers: project_rsnek

Maniphest Tasks: T42

Differential Revision: https://dillon.phacility.com/D2